+++ packages/iov-bcp/src/atomicswaphelpers.spec.ts
@@ -10,6 +10,13 @@ describe("AtomicSwapHelpers", () => {
     it("works", async () => {
       const preimage = await AtomicSwapHelpers.createPreimage();
       expect(preimage.length).toEqual(32);
+      expect(new Set(preimage).size).toBeGreaterThanOrEqual(25);
+    });
+
+    it("created different results", async () => {
+      const value1 = await AtomicSwapHelpers.createPreimage();
+      const value2 = await AtomicSwapHelpers.createPreimage();
+      expect(value1).not.toEqual(value2);
     });
   });
 
@@ -19,6 +26,20 @@ describe("AtomicSwapHelpers", () => {
       const preimage = new Uint8Array([]) as Preimage;
       const hash = AtomicSwapHelpers.hashPreimage(preimage);
       expect(hash).toEqual(fromHex("e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"));
+    });
+  });
+
+  describe("createId", () => {
+    it("works", async () => {
+      const id = await AtomicSwapHelpers.createId();
+      expect(id.length).toEqual(32);
+      expect(new Set(id).size).toBeGreaterThanOrEqual(25);
+    });
+
+    it("created different results", async () => {
+      const value1 = await AtomicSwapHelpers.createId();
+      const value2 = await AtomicSwapHelpers.createId();
+      expect(value1).not.toEqual(value2);
     });
   });
 });